--- src/models/index.d.ts
@@ -1,8 +1,0 @@
-interface IConfig {
-  azure: {
-    tennant: string
-    appId: string
-    issuer: string
-    jwk_endpoint: string
-  };
-}
