+++ src/services/JWTService.ts
@@ -2,6 +2,7 @@ import * as JWT from "jsonwebtoken";
 import * as http from "request-promise";
 import AuthorizationError from "../models/exceptions/AuthorizationError";
 import {ALLOWEDROLES, ERRORMESSAGES} from "../assets/enum";
+import {IConfig} from "../utils/IConfig";
 
 class JWTService {
 
@@ -10,7 +11,7 @@ class JWTService {
    * @param token
    * @param config
    */
-  public async verify(token: string, config: IConfig): Promise<any> {
+  public async verify(token: string, config: IConfig): Promise<string | object> {
     const decodedToken: any = JWT.decode(token, {complete: true});
 
     // Check if config is valid
